Add CoreWingMovement test module
Move WingMovement tests to a Core-level test module so they directly
test CoreWingMovement behavior. Keep a slim WingMovement test module
that verifies the public subclass relationship, the stricter type check
on wing_cross_section_movements, and inherited behavior. Duplicate the
nine WingMovement fixtures that consumer tests need since
AirplaneMovement enforces the stricter public type.
