+++ tests/unit/__init__.py
@@ -23,6 +23,9 @@ This package contains the following modules:
     test_core_wing_cross_section_movement.py: This module contains classes to test
     CoreWingCrossSectionMovements.
 
+    test_core_wing_movement.py: This module contains classes to test
+    CoreWingMovements.
+
     test_horseshoe_vortex.py: This module contains classes to test HorseshoeVortices.
 
     test_line_vortex.py: This module contains classes to test LineVortices.
@@ -75,6 +78,7 @@ import tests.unit.test_airplane
 import tests.unit.test_airplane_movement
 import tests.unit.test_core
 import tests.unit.test_core_wing_cross_section_movement
+import tests.unit.test_core_wing_movement
 import tests.unit.test_horseshoe_vortex
 import tests.unit.test_line_vortex
 import tests.unit.test_movement
