+++ tests/unit/fixtures/wing_movement_fixtures.py
@@ -8,15 +8,15 @@ from . import geometry_fixtures, wing_cross_section_movement_fixtures
 
 
 def make_static_wing_movement_fixture():
-    """This method makes a fixture that is a WingMovement with all parameters
-    zero (no movement).
+    """This method makes a fixture that is a WingMovement with all parameters zero
+    (no movement).
 
     :return static_wing_movement_fixture: WingMovement
         This is the WingMovement with no movement.
     """
     # Initialize the constructing fixtures.
     base_wing = geometry_fixtures.make_origin_wing_fixture()
-    wcs_movements = [
+    wing_cross_section_movements = [
         wing_cross_section_movement_fixtures.make_static_wing_cross_section_movement_fixture(),
         wing_cross_section_movement_fixtures.make_static_tip_wing_cross_section_movement_fixture(),
     ]
@@ -24,7 +24,7 @@ def make_static_wing_movement_fixture():
     # Create the static WingMovement.
     static_wing_movement_fixture = ps.movements.wing_movement.WingMovement(
         base_wing=base_wing,
-        wing_cross_section_movements=wcs_movements,
+        wing_cross_section_movements=wing_cross_section_movements,
         ampLer_Gs_Cgs=(0.0, 0.0, 0.0),
         periodLer_Gs_Cgs=(0.0, 0.0, 0.0),
         spacingLer_Gs_Cgs=("sine", "sine", "sine"),
@@ -48,7 +48,7 @@ def make_basic_wing_movement_fixture():
     """
     # Initialize the constructing fixtures.
     base_wing = geometry_fixtures.make_origin_wing_fixture()
-    wcs_movements = [
+    wing_cross_section_movements = [
         wing_cross_section_movement_fixtures.make_static_wing_cross_section_movement_fixture(),
         wing_cross_section_movement_fixtures.make_basic_wing_cross_section_movement_fixture(),
     ]
@@ -56,7 +56,7 @@ def make_basic_wing_movement_fixture():
     # Create the basic WingMovement.
     basic_wing_movement_fixture = ps.movements.wing_movement.WingMovement(
         base_wing=base_wing,
-        wing_cross_section_movements=wcs_movements,
+        wing_cross_section_movements=wing_cross_section_movements,
         ampLer_Gs_Cgs=(0.1, 0.05, 0.08),
         periodLer_Gs_Cgs=(2.0, 2.0, 2.0),
         spacingLer_Gs_Cgs=("sine", "sine", "sine"),
@@ -80,15 +80,15 @@ def make_sine_spacing_Ler_wing_movement_fixture():
     """
     # Initialize the constructing fixtures.
     base_wing = geometry_fixtures.make_origin_wing_fixture()
-    wcs_movements = [
+    wing_cross_section_movements = [
         wing_cross_section_movement_fixtures.make_static_wing_cross_section_movement_fixture(),
         wing_cross_section_movement_fixtures.make_static_tip_wing_cross_section_movement_fixture(),
     ]
 
-    # Create the WingMovement with sine spacing for Ler_Gs_Cgs.
+    # Create the sine-spacing-Ler WingMovement.
     sine_spacing_Ler_wing_movement_fixture = ps.movements.wing_movement.WingMovement(
         base_wing=base_wing,
-        wing_cross_section_movements=wcs_movements,
+        wing_cross_section_movements=wing_cross_section_movements,
         ampLer_Gs_Cgs=(0.2, 0.0, 0.0),
         periodLer_Gs_Cgs=(1.0, 0.0, 0.0),
         spacingLer_Gs_Cgs=("sine", "sine", "sine"),
@@ -103,212 +103,15 @@ def make_sine_spacing_Ler_wing_movement_fixture():
     return sine_spacing_Ler_wing_movement_fixture
 
 
-def make_uniform_spacing_Ler_wing_movement_fixture():
-    """This method makes a fixture that is a WingMovement with uniform spacing for
-    Ler_Gs_Cgs.
-
-    :return uniform_spacing_Ler_wing_movement_fixture: WingMovement
-        This is the WingMovement with uniform spacing for Ler_Gs_Cgs.
-    """
-    # Initialize the constructing fixtures.
-    base_wing = geometry_fixtures.make_origin_wing_fixture()
-    wcs_movements = [
-        wing_cross_section_movement_fixtures.make_static_wing_cross_section_movement_fixture(),
-        wing_cross_section_movement_fixtures.make_static_tip_wing_cross_section_movement_fixture(),
-    ]
-
-    # Create the WingMovement with uniform spacing for Ler_Gs_Cgs.
-    uniform_spacing_Ler_wing_movement_fixture = ps.movements.wing_movement.WingMovement(
-        base_wing=base_wing,
-        wing_cross_section_movements=wcs_movements,
-        ampLer_Gs_Cgs=(0.2, 0.0, 0.0),
-        periodLer_Gs_Cgs=(1.0, 0.0, 0.0),
-        spacingLer_Gs_Cgs=("uniform", "uniform", "uniform"),
-        phaseLer_Gs_Cgs=(0.0, 0.0, 0.0),
-        ampAngles_Gs_to_Wn_ixyz=(0.0, 0.0, 0.0),
-        periodAngles_Gs_to_Wn_ixyz=(0.0, 0.0, 0.0),
-        spacingAngles_Gs_to_Wn_ixyz=("sine", "sine", "sine"),
-        phaseAngles_Gs_to_Wn_ixyz=(0.0, 0.0, 0.0),
-    )
-
-    # Return the WingMovement fixture.
-    return uniform_spacing_Ler_wing_movement_fixture
-
-
-def make_mixed_spacing_Ler_wing_movement_fixture():
-    """This method makes a fixture that is a WingMovement with mixed spacing for
-    Ler_Gs_Cgs.
-
-    :return mixed_spacing_Ler_wing_movement_fixture: WingMovement
-        This is the WingMovement with mixed spacing for Ler_Gs_Cgs.
-    """
-    # Initialize the constructing fixtures.
-    base_wing = geometry_fixtures.make_origin_wing_fixture()
-    wcs_movements = [
-        wing_cross_section_movement_fixtures.make_static_wing_cross_section_movement_fixture(),
-        wing_cross_section_movement_fixtures.make_static_tip_wing_cross_section_movement_fixture(),
-    ]
-
-    # Create the WingMovement with mixed spacing for Ler_Gs_Cgs.
-    mixed_spacing_Ler_wing_movement_fixture = ps.movements.wing_movement.WingMovement(
-        base_wing=base_wing,
-        wing_cross_section_movements=wcs_movements,
-        ampLer_Gs_Cgs=(0.2, 0.15, 0.1),
-        periodLer_Gs_Cgs=(1.0, 1.0, 1.0),
-        spacingLer_Gs_Cgs=("sine", "uniform", "sine"),
-        phaseLer_Gs_Cgs=(0.0, 0.0, 0.0),
-        ampAngles_Gs_to_Wn_ixyz=(0.0, 0.0, 0.0),
-        periodAngles_Gs_to_Wn_ixyz=(0.0, 0.0, 0.0),
-        spacingAngles_Gs_to_Wn_ixyz=("sine", "sine", "sine"),
-        phaseAngles_Gs_to_Wn_ixyz=(0.0, 0.0, 0.0),
-    )
-
-    # Return the WingMovement fixture.
-    return mixed_spacing_Ler_wing_movement_fixture
-
-
-def make_sine_spacing_angles_wing_movement_fixture():
-    """This method makes a fixture that is a WingMovement with sine spacing for
-    angles_Gs_to_Wn_ixyz.
-
-    :return sine_spacing_angles_wing_movement_fixture: WingMovement
-        This is the WingMovement with sine spacing for angles_Gs_to_Wn_ixyz.
-    """
-    # Initialize the constructing fixtures.
-    base_wing = geometry_fixtures.make_origin_wing_fixture()
-    wcs_movements = [
-        wing_cross_section_movement_fixtures.make_static_wing_cross_section_movement_fixture(),
-        wing_cross_section_movement_fixtures.make_static_tip_wing_cross_section_movement_fixture(),
-    ]
-
-    # Create the WingMovement with sine spacing for angles_Gs_to_Wn_ixyz.
-    sine_spacing_angles_wing_movement_fixture = ps.movements.wing_movement.WingMovement(
-        base_wing=base_wing,
-        wing_cross_section_movements=wcs_movements,
-        ampLer_Gs_Cgs=(0.0, 0.0, 0.0),
-        periodLer_Gs_Cgs=(0.0, 0.0, 0.0),
-        spacingLer_Gs_Cgs=("sine", "sine", "sine"),
-        phaseLer_Gs_Cgs=(0.0, 0.0, 0.0),
-        ampAngles_Gs_to_Wn_ixyz=(10.0, 0.0, 0.0),
-        periodAngles_Gs_to_Wn_ixyz=(1.0, 0.0, 0.0),
-        spacingAngles_Gs_to_Wn_ixyz=("sine", "sine", "sine"),
-        phaseAngles_Gs_to_Wn_ixyz=(0.0, 0.0, 0.0),
-    )
-
-    # Return the WingMovement fixture.
-    return sine_spacing_angles_wing_movement_fixture
-
-
-def make_uniform_spacing_angles_wing_movement_fixture():
-    """This method makes a fixture that is a WingMovement with uniform spacing for
-    angles_Gs_to_Wn_ixyz.
-
-    :return uniform_spacing_angles_wing_movement_fixture: WingMovement
-        This is the WingMovement with uniform spacing for angles_Gs_to_Wn_ixyz.
-    """
-    # Initialize the constructing fixtures.
-    base_wing = geometry_fixtures.make_origin_wing_fixture()
-    wcs_movements = [
-        wing_cross_section_movement_fixtures.make_static_wing_cross_section_movement_fixture(),
-        wing_cross_section_movement_fixtures.make_static_tip_wing_cross_section_movement_fixture(),
-    ]
-
-    # Create the WingMovement with uniform spacing for angles_Gs_to_Wn_ixyz.
-    uniform_spacing_angles_wing_movement_fixture = (
-        ps.movements.wing_movement.WingMovement(
-            base_wing=base_wing,
-            wing_cross_section_movements=wcs_movements,
-            ampLer_Gs_Cgs=(0.0, 0.0, 0.0),
-            periodLer_Gs_Cgs=(0.0, 0.0, 0.0),
-            spacingLer_Gs_Cgs=("sine", "sine", "sine"),
-            phaseLer_Gs_Cgs=(0.0, 0.0, 0.0),
-            ampAngles_Gs_to_Wn_ixyz=(10.0, 0.0, 0.0),
-            periodAngles_Gs_to_Wn_ixyz=(1.0, 0.0, 0.0),
-            spacingAngles_Gs_to_Wn_ixyz=("uniform", "uniform", "uniform"),
-            phaseAngles_Gs_to_Wn_ixyz=(0.0, 0.0, 0.0),
-        )
-    )
-
-    # Return the WingMovement fixture.
-    return uniform_spacing_angles_wing_movement_fixture
-
-
-def make_mixed_spacing_angles_wing_movement_fixture():
-    """This method makes a fixture that is a WingMovement with mixed spacing for
-    angles_Gs_to_Wn_ixyz.
-
-    :return mixed_spacing_angles_wing_movement_fixture: WingMovement
-        This is the WingMovement with mixed spacing for angles_Gs_to_Wn_ixyz.
-    """
-    # Initialize the constructing fixtures.
-    base_wing = geometry_fixtures.make_origin_wing_fixture()
-    wcs_movements = [
-        wing_cross_section_movement_fixtures.make_static_wing_cross_section_movement_fixture(),
-        wing_cross_section_movement_fixtures.make_static_tip_wing_cross_section_movement_fixture(),
-    ]
-
-    # Create the WingMovement with mixed spacing for angles_Gs_to_Wn_ixyz.
-    mixed_spacing_angles_wing_movement_fixture = (
-        ps.movements.wing_movement.WingMovement(
-            base_wing=base_wing,
-            wing_cross_section_movements=wcs_movements,
-            ampLer_Gs_Cgs=(0.0, 0.0, 0.0),
-            periodLer_Gs_Cgs=(0.0, 0.0, 0.0),
-            spacingLer_Gs_Cgs=("sine", "sine", "sine"),
-            phaseLer_Gs_Cgs=(0.0, 0.0, 0.0),
-            ampAngles_Gs_to_Wn_ixyz=(10.0, 15.0, 8.0),
-            periodAngles_Gs_to_Wn_ixyz=(1.0, 1.0, 1.0),
-            spacingAngles_Gs_to_Wn_ixyz=("sine", "uniform", "sine"),
-            phaseAngles_Gs_to_Wn_ixyz=(0.0, 0.0, 0.0),
-        )
-    )
-
-    # Return the WingMovement fixture.
-    return mixed_spacing_angles_wing_movement_fixture
-
-
-def make_Ler_only_wing_movement_fixture():
-    """This method makes a fixture that is a WingMovement where only Ler_Gs_Cgs
-    moves.
-
-    :return Ler_only_wing_movement_fixture: WingMovement
-        This is the WingMovement with only Ler_Gs_Cgs movement.
-    """
-    # Initialize the constructing fixtures.
-    base_wing = geometry_fixtures.make_origin_wing_fixture()
-    wcs_movements = [
-        wing_cross_section_movement_fixtures.make_static_wing_cross_section_movement_fixture(),
-        wing_cross_section_movement_fixtures.make_static_tip_wing_cross_section_movement_fixture(),
-    ]
-
-    # Create the Ler-only WingMovement.
-    Ler_only_wing_movement_fixture = ps.movements.wing_movement.WingMovement(
-        base_wing=base_wing,
-        wing_cross_section_movements=wcs_movements,
-        ampLer_Gs_Cgs=(0.15, 0.1, 0.08),
-        periodLer_Gs_Cgs=(1.5, 1.5, 1.5),
-        spacingLer_Gs_Cgs=("sine", "sine", "sine"),
-        phaseLer_Gs_Cgs=(0.0, 0.0, 0.0),
-        ampAngles_Gs_to_Wn_ixyz=(0.0, 0.0, 0.0),
-        periodAngles_Gs_to_Wn_ixyz=(0.0, 0.0, 0.0),
-        spacingAngles_Gs_to_Wn_ixyz=("sine", "sine", "sine"),
-        phaseAngles_Gs_to_Wn_ixyz=(0.0, 0.0, 0.0),
-    )
-
-    # Return the WingMovement fixture.
-    return Ler_only_wing_movement_fixture
-
-
 def make_angles_only_wing_movement_fixture():
-    """This method makes a fixture that is a WingMovement where only
-    angles_Gs_to_Wn_ixyz moves.
+    """This method makes a fixture that is a WingMovement with only angle movement.
 
     :return angles_only_wing_movement_fixture: WingMovement
-        This is the WingMovement with only angles_Gs_to_Wn_ixyz movement.
+        This is the WingMovement with only angle movement.
     """
     # Initialize the constructing fixtures.
     base_wing = geometry_fixtures.make_origin_wing_fixture()
-    wcs_movements = [
+    wing_cross_section_movements = [
         wing_cross_section_movement_fixtures.make_static_wing_cross_section_movement_fixture(),
         wing_cross_section_movement_fixtures.make_static_tip_wing_cross_section_movement_fixture(),
     ]
@@ -316,7 +119,7 @@ def make_angles_only_wing_movement_fixture():
     # Create the angles-only WingMovement.
     angles_only_wing_movement_fixture = ps.movements.wing_movement.WingMovement(
         base_wing=base_wing,
-        wing_cross_section_movements=wcs_movements,
+        wing_cross_section_movements=wing_cross_section_movements,
         ampLer_Gs_Cgs=(0.0, 0.0, 0.0),
         periodLer_Gs_Cgs=(0.0, 0.0, 0.0),
         spacingLer_Gs_Cgs=("sine", "sine", "sine"),
@@ -331,80 +134,16 @@ def make_angles_only_wing_movement_fixture():
     return angles_only_wing_movement_fixture
 
 
-def make_phase_offset_Ler_wing_movement_fixture():
-    """This method makes a fixture that is a WingMovement with non-zero phase offset
-    for Ler_Gs_Cgs.
-
-    :return phase_offset_Ler_wing_movement_fixture: WingMovement
-        This is the WingMovement with phase offset for Ler_Gs_Cgs.
-    """
-    # Initialize the constructing fixtures.
-    base_wing = geometry_fixtures.make_origin_wing_fixture()
-    wcs_movements = [
-        wing_cross_section_movement_fixtures.make_static_wing_cross_section_movement_fixture(),
-        wing_cross_section_movement_fixtures.make_static_tip_wing_cross_section_movement_fixture(),
-    ]
-
-    # Create the phase-offset WingMovement.
-    phase_offset_Ler_wing_movement_fixture = ps.movements.wing_movement.WingMovement(
-        base_wing=base_wing,
-        wing_cross_section_movements=wcs_movements,
-        ampLer_Gs_Cgs=(0.1, 0.08, 0.06),
-        periodLer_Gs_Cgs=(1.0, 1.0, 1.0),
-        spacingLer_Gs_Cgs=("sine", "sine", "sine"),
-        phaseLer_Gs_Cgs=(90.0, -45.0, 60.0),
-        ampAngles_Gs_to_Wn_ixyz=(0.0, 0.0, 0.0),
-        periodAngles_Gs_to_Wn_ixyz=(0.0, 0.0, 0.0),
-        spacingAngles_Gs_to_Wn_ixyz=("sine", "sine", "sine"),
-        phaseAngles_Gs_to_Wn_ixyz=(0.0, 0.0, 0.0),
-    )
-
-    # Return the WingMovement fixture.
-    return phase_offset_Ler_wing_movement_fixture
-
-
-def make_phase_offset_angles_wing_movement_fixture():
-    """This method makes a fixture that is a WingMovement with non-zero phase offset
-    for angles_Gs_to_Wn_ixyz.
-
-    :return phase_offset_angles_wing_movement_fixture: WingMovement
-        This is the WingMovement with phase offset for angles_Gs_to_Wn_ixyz.
-    """
-    # Initialize the constructing fixtures.
-    base_wing = geometry_fixtures.make_origin_wing_fixture()
-    wcs_movements = [
-        wing_cross_section_movement_fixtures.make_static_wing_cross_section_movement_fixture(),
-        wing_cross_section_movement_fixtures.make_static_tip_wing_cross_section_movement_fixture(),
-    ]
-
-    # Create the phase-offset WingMovement.
-    phase_offset_angles_wing_movement_fixture = ps.movements.wing_movement.WingMovement(
-        base_wing=base_wing,
-        wing_cross_section_movements=wcs_movements,
-        ampLer_Gs_Cgs=(0.0, 0.0, 0.0),
-        periodLer_Gs_Cgs=(0.0, 0.0, 0.0),
-        spacingLer_Gs_Cgs=("sine", "sine", "sine"),
-        phaseLer_Gs_Cgs=(0.0, 0.0, 0.0),
-        ampAngles_Gs_to_Wn_ixyz=(10.0, 12.0, 8.0),
-        periodAngles_Gs_to_Wn_ixyz=(1.0, 1.0, 1.0),
-        spacingAngles_Gs_to_Wn_ixyz=("sine", "sine", "sine"),
-        phaseAngles_Gs_to_Wn_ixyz=(45.0, 90.0, -30.0),
-    )
-
-    # Return the WingMovement fixture.
-    return phase_offset_angles_wing_movement_fixture
-
-
 def make_multiple_periods_wing_movement_fixture():
-    """This method makes a fixture that is a WingMovement with different periods
-    for different dimensions.
+    """This method makes a fixture that is a WingMovement with different periods for
+    different dimensions.
 
     :return multiple_periods_wing_movement_fixture: WingMovement
         This is the WingMovement with different periods.
     """
     # Initialize the constructing fixtures.
     base_wing = geometry_fixtures.make_origin_wing_fixture()
-    wcs_movements = [
+    wing_cross_section_movements = [
         wing_cross_section_movement_fixtures.make_static_wing_cross_section_movement_fixture(),
         wing_cross_section_movement_fixtures.make_multiple_periods_wing_cross_section_movement_fixture(),
     ]
@@ -412,7 +151,7 @@ def make_multiple_periods_wing_movement_fixture():
     # Create the multiple-periods WingMovement.
     multiple_periods_wing_movement_fixture = ps.movements.wing_movement.WingMovement(
         base_wing=base_wing,
-        wing_cross_section_movements=wcs_movements,
+        wing_cross_section_movements=wing_cross_section_movements,
         ampLer_Gs_Cgs=(0.1, 0.08, 0.06),
         periodLer_Gs_Cgs=(1.0, 2.0, 3.0),
         spacingLer_Gs_Cgs=("sine", "sine", "sine"),
@@ -427,106 +166,6 @@ def make_multiple_periods_wing_movement_fixture():
     return multiple_periods_wing_movement_fixture
 
 
-def make_custom_spacing_Ler_wing_movement_fixture():
-    """This method makes a fixture that is a WingMovement with a custom spacing
-    function for Ler_Gs_Cgs.
-
-    :return custom_spacing_Ler_wing_movement_fixture: WingMovement
-        This is the WingMovement with custom spacing for Ler_Gs_Cgs.
-    """
-    # Initialize the constructing fixtures.
-    base_wing = geometry_fixtures.make_origin_wing_fixture()
-    wcs_movements = [
-        wing_cross_section_movement_fixtures.make_static_wing_cross_section_movement_fixture(),
-        wing_cross_section_movement_fixtures.make_static_tip_wing_cross_section_movement_fixture(),
-    ]
-
-    # Define a custom harmonic spacing function.
-    def custom_harmonic(x):
-        """Custom harmonic spacing function: normalized combination of harmonics.
-
-        This function satisfies all requirements: starts at 0, returns to 0 at
-        2*pi, has zero mean, has amplitude of 1, and is periodic.
-
-        :param x: (N,) ndarray of floats
-            The input angles in radians.
-
-        :return: (N,) ndarray of floats
-            The output values.
-        """
-        return (3.0 / (2.0 * np.sqrt(2.0))) * (
-            np.sin(x) + (1.0 / 3.0) * np.sin(3.0 * x)
-        )
-
-    # Create the custom-spacing WingMovement.
-    custom_spacing_Ler_wing_movement_fixture = ps.movements.wing_movement.WingMovement(
-        base_wing=base_wing,
-        wing_cross_section_movements=wcs_movements,
-        ampLer_Gs_Cgs=(0.15, 0.0, 0.0),
-        periodLer_Gs_Cgs=(1.0, 0.0, 0.0),
-        spacingLer_Gs_Cgs=(custom_harmonic, "sine", "sine"),
-        phaseLer_Gs_Cgs=(0.0, 0.0, 0.0),
-        ampAngles_Gs_to_Wn_ixyz=(0.0, 0.0, 0.0),
-        periodAngles_Gs_to_Wn_ixyz=(0.0, 0.0, 0.0),
-        spacingAngles_Gs_to_Wn_ixyz=("sine", "sine", "sine"),
-        phaseAngles_Gs_to_Wn_ixyz=(0.0, 0.0, 0.0),
-    )
-
-    # Return the WingMovement fixture.
-    return custom_spacing_Ler_wing_movement_fixture
-
-
-def make_custom_spacing_angles_wing_movement_fixture():
-    """This method makes a fixture that is a WingMovement with a custom spacing
-    function for angles_Gs_to_Wn_ixyz.
-
-    :return custom_spacing_angles_wing_movement_fixture: WingMovement
-        This is the WingMovement with custom spacing for angles_Gs_to_Wn_ixyz.
-    """
-    # Initialize the constructing fixtures.
-    base_wing = geometry_fixtures.make_origin_wing_fixture()
-    wcs_movements = [
-        wing_cross_section_movement_fixtures.make_static_wing_cross_section_movement_fixture(),
-        wing_cross_section_movement_fixtures.make_static_tip_wing_cross_section_movement_fixture(),
-    ]
-
-    # Define a custom harmonic spacing function.
-    def custom_harmonic(x):
-        """Custom harmonic spacing function: normalized combination of harmonics.
-
-        This function satisfies all requirements: starts at 0, returns to 0 at
-        2*pi, has zero mean, has amplitude of 1, and is periodic.
-
-        :param x: (N,) ndarray of floats
-            The input angles in radians.
-
-        :return: (N,) ndarray of floats
-            The output values.
-        """
-        return (3.0 / (2.0 * np.sqrt(2.0))) * (
-            np.sin(x) + (1.0 / 3.0) * np.sin(3.0 * x)
-        )
-
-    # Create the custom-spacing WingMovement.
-    custom_spacing_angles_wing_movement_fixture = (
-        ps.movements.wing_movement.WingMovement(
-            base_wing=base_wing,
-            wing_cross_section_movements=wcs_movements,
-            ampLer_Gs_Cgs=(0.0, 0.0, 0.0),
-            periodLer_Gs_Cgs=(0.0, 0.0, 0.0),
-            spacingLer_Gs_Cgs=("sine", "sine", "sine"),
-            phaseLer_Gs_Cgs=(0.0, 0.0, 0.0),
-            ampAngles_Gs_to_Wn_ixyz=(10.0, 0.0, 0.0),
-            periodAngles_Gs_to_Wn_ixyz=(1.0, 0.0, 0.0),
-            spacingAngles_Gs_to_Wn_ixyz=(custom_harmonic, "sine", "sine"),
-            phaseAngles_Gs_to_Wn_ixyz=(0.0, 0.0, 0.0),
-        )
-    )
-
-    # Return the WingMovement fixture.
-    return custom_spacing_angles_wing_movement_fixture
-
-
 def make_mixed_custom_and_standard_spacing_wing_movement_fixture():
     """This method makes a fixture that is a WingMovement with mixed custom and
     standard spacing functions.
@@ -536,7 +175,7 @@ def make_mixed_custom_and_standard_spacing_wing_movement_fixture():
     """
     # Initialize the constructing fixtures.
     base_wing = geometry_fixtures.make_origin_wing_fixture()
-    wcs_movements = [
+    wing_cross_section_movements = [
         wing_cross_section_movement_fixtures.make_static_wing_cross_section_movement_fixture(),
         wing_cross_section_movement_fixtures.make_mixed_custom_and_standard_spacing_wing_cross_section_movement_fixture(),
     ]
@@ -559,7 +198,7 @@ def make_mixed_custom_and_standard_spacing_wing_movement_fixture():
     mixed_custom_and_standard_spacing_wing_movement_fixture = (
         ps.movements.wing_movement.WingMovement(
             base_wing=base_wing,
-            wing_cross_section_movements=wcs_movements,
+            wing_cross_section_movements=wing_cross_section_movements,
             ampLer_Gs_Cgs=(0.1, 0.08, 0.06),
             periodLer_Gs_Cgs=(1.0, 1.0, 1.0),
             spacingLer_Gs_Cgs=(custom_harmonic, "uniform", "sine"),
@@ -575,65 +214,26 @@ def make_mixed_custom_and_standard_spacing_wing_movement_fixture():
     return mixed_custom_and_standard_spacing_wing_movement_fixture
 
 
-def make_rotation_point_offset_wing_movement_fixture():
-    """This method makes a fixture that is a WingMovement with a non zero rotation
-    point offset.
-
-    :return rotation_point_offset_wing_movement_fixture: WingMovement
-        This is the WingMovement with a non zero rotationPointOffset_Gs_Ler.
-    """
-    # Initialize the constructing fixtures.
-    base_wing = geometry_fixtures.make_origin_wing_fixture()
-    wcs_movements = [
-        wing_cross_section_movement_fixtures.make_static_wing_cross_section_movement_fixture(),
-        wing_cross_section_movement_fixtures.make_static_tip_wing_cross_section_movement_fixture(),
-    ]
-
-    # Create the WingMovement with rotation point offset.
-    # The offset is in y direction (0.0, 0.5, 0.0), and we rotate about the x axis.
-    # This causes the wing to trace an arc in the yz plane as it rotates.
-    rotation_point_offset_wing_movement_fixture = (
-        ps.movements.wing_movement.WingMovement(
-            base_wing=base_wing,
-            wing_cross_section_movements=wcs_movements,
-            ampLer_Gs_Cgs=(0.0, 0.0, 0.0),
-            periodLer_Gs_Cgs=(0.0, 0.0, 0.0),
-            spacingLer_Gs_Cgs=("sine", "sine", "sine"),
-            phaseLer_Gs_Cgs=(0.0, 0.0, 0.0),
-            ampAngles_Gs_to_Wn_ixyz=(10.0, 0.0, 0.0),
-            periodAngles_Gs_to_Wn_ixyz=(1.0, 0.0, 0.0),
-            spacingAngles_Gs_to_Wn_ixyz=("sine", "sine", "sine"),
-            phaseAngles_Gs_to_Wn_ixyz=(0.0, 0.0, 0.0),
-            rotationPointOffset_Gs_Ler=(0.0, 0.5, 0.0),
-        )
-    )
-
-    # Return the WingMovement fixture.
-    return rotation_point_offset_wing_movement_fixture
-
-
 def make_periodic_geometry_wing_movement_fixture():
-    """This method makes a fixture that is a WingMovement with periodic geometry motion
-    suitable for testing the variable geometry optimization.
-
-    The fixture uses a 0.1s period which aligns well with common delta_time values
-    like 0.01s (10 steps per period) and 0.02s (5 steps per period).
+    """This method makes a fixture that is a WingMovement with periodic geometry
+    motion, suitable for testing the variable geometry optimization. The fixture
+    uses a 0.1s period which aligns well with common delta_time values like 0.01s
+    (10 steps per period) and 0.02s (5 steps per period).
 
     :return periodic_geometry_wing_movement_fixture: WingMovement
         This is the WingMovement with periodic geometry motion.
     """
     # Initialize the constructing fixtures.
     base_wing = geometry_fixtures.make_origin_wing_fixture()
-    wcs_movements = [
+    wing_cross_section_movements = [
         wing_cross_section_movement_fixtures.make_static_wing_cross_section_movement_fixture(),
         wing_cross_section_movement_fixtures.make_static_tip_wing_cross_section_movement_fixture(),
     ]
 
-    # Create the periodic geometry WingMovement.
-    # Use a 0.1s period for angular motion (plunging wing motion).
+    # Create the periodic-geometry WingMovement.
     periodic_geometry_wing_movement_fixture = ps.movements.wing_movement.WingMovement(
         base_wing=base_wing,
-        wing_cross_section_movements=wcs_movements,
+        wing_cross_section_movements=wing_cross_section_movements,
         ampLer_Gs_Cgs=(0.0, 0.0, 0.0),
         periodLer_Gs_Cgs=(0.0, 0.0, 0.0),
         spacingLer_Gs_Cgs=("sine", "sine", "sine"),
@@ -649,20 +249,23 @@ def make_periodic_geometry_wing_movement_fixture():
 
 
 def make_2_chordwise_panels_wing_movement_fixture():
-    """This method makes a fixture that is a WingMovement for a Wing with
-    2 chordwise panels.
+    """This method makes a fixture that is a WingMovement for a Wing with 2
+    chordwise panels.
 
-    :return: WingMovement for a Wing with 2 chordwise panels.
+    :return fixture: WingMovement
+        This is the WingMovement for a Wing with 2 chordwise panels.
     """
+    # Initialize the constructing fixtures.
     base_wing = geometry_fixtures.make_wing_with_2_chordwise_panels()
-    wcs_movements = [
+    wing_cross_section_movements = [
         wing_cross_section_movement_fixtures.make_static_wing_cross_section_movement_fixture(),
         wing_cross_section_movement_fixtures.make_static_tip_wing_cross_section_movement_fixture(),
     ]
 
+    # Create the WingMovement for a Wing with 2 chordwise panels.
     fixture = ps.movements.wing_movement.WingMovement(
         base_wing=base_wing,
-        wing_cross_section_movements=wcs_movements,
+        wing_cross_section_movements=wing_cross_section_movements,
         ampLer_Gs_Cgs=(0.0, 0.0, 0.0),
         periodLer_Gs_Cgs=(0.0, 0.0, 0.0),
         spacingLer_Gs_Cgs=("sine", "sine", "sine"),
@@ -673,24 +276,28 @@ def make_2_chordwise_panels_wing_movement_fixture():
         phaseAngles_Gs_to_Wn_ixyz=(0.0, 0.0, 0.0),
     )
 
+    # Return the WingMovement fixture.
     return fixture
 
 
 def make_3_chordwise_panels_wing_movement_fixture():
-    """This method makes a fixture that is a WingMovement for a Wing with
-    3 chordwise panels.
+    """This method makes a fixture that is a WingMovement for a Wing with 3
+    chordwise panels.
 
-    :return: WingMovement for a Wing with 3 chordwise panels.
+    :return fixture: WingMovement
+        This is the WingMovement for a Wing with 3 chordwise panels.
     """
+    # Initialize the constructing fixtures.
     base_wing = geometry_fixtures.make_wing_with_3_chordwise_panels()
-    wcs_movements = [
+    wing_cross_section_movements = [
         wing_cross_section_movement_fixtures.make_static_wing_cross_section_movement_fixture(),
         wing_cross_section_movement_fixtures.make_static_tip_wing_cross_section_movement_fixture(),
     ]
 
+    # Create the WingMovement for a Wing with 3 chordwise panels.
     fixture = ps.movements.wing_movement.WingMovement(
         base_wing=base_wing,
-        wing_cross_section_movements=wcs_movements,
+        wing_cross_section_movements=wing_cross_section_movements,
         ampLer_Gs_Cgs=(0.0, 0.0, 0.0),
         periodLer_Gs_Cgs=(0.0, 0.0, 0.0),
         spacingLer_Gs_Cgs=("sine", "sine", "sine"),
@@ -701,4 +308,5 @@ def make_3_chordwise_panels_wing_movement_fixture():
         phaseAngles_Gs_to_Wn_ixyz=(0.0, 0.0, 0.0),
     )
 
+    # Return the WingMovement fixture.
     return fixture
